% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{aprean3}
\alias{aprean3}
\alias{aprean3-package}
\title{Datasets from Draper and Smith "Applied Regression Analysis" (3rd ed.)}
\source{
Draper, N.R., Smith, H. (1998) Applied Regression Analysis, 3rd
ed., New York: Wiley
}
\description{
This package is a unofficial companion to the textbook "Applied
Regression Analysis" by N.R. Draper and H. Smith (3rd ed.).
}
\details{
It includes all the datasets used in the book, both for easy reproducibility
(assuming a didactic point of view) and algorithms benchmarking purposes.

Dataset names are prefixed by 'ds' and lowercased; furthermore '-' and
'.' are replaced by '_'. Finally variables have lower-case name.
}

