% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds01a}
\alias{ds01a}
\title{Dataset for Appendix A, Chapter 01}
\format{A data.frame with 25 rows and 10 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{x6}{}
\item{x7}{}
\item{x8}{}
\item{x9}{}
\item{x10}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds01a
}
\description{
Dataset for Appendix A, Chapter 01
}
\examples{
ds01a
}
\keyword{datasets}

