% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds02_1}
\alias{ds02_1}
\title{02-1 data}
\format{A data.frame with 23 rows and 2 variables:
\describe{
\item{y}{}
\item{x}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds02_1
}
\description{
02-1 data
}
\examples{
ds02_1
}
\keyword{datasets}

