% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds03_2}
\alias{ds03_2}
\title{03-2 data}
\format{A data.frame with 27 rows and 5 variables:
\describe{
\item{X1}{}
\item{X12.60}{}
\item{X2.449}{}
\item{X11.39}{}
\item{X2.242}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds03_2
}
\description{
ds03_2 data
}
\examples{
ds03_2
}
\keyword{datasets}

