% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds03_3}
\alias{ds03_3}
\title{03-3 data}
\format{A data.frame with 20 rows and 3 variables:
\describe{
\item{id}{}
\item{x}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds03_3
}
\description{
ds03_3 data
}
\examples{
ds03_3
}
\keyword{datasets}

