% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds06b}
\alias{ds06b}
\title{Dataset for Appendix B, Chapter 06}
\format{A data.frame with 70 rows and 4 variables:
\describe{
\item{set}{}
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds06b
}
\description{
Dataset for Appendix B, Chapter 06
}
\examples{
ds06b
}
\keyword{datasets}

