% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds08_1}
\alias{ds08_1}
\title{Dataset for Table 08.1}
\format{A data.frame with 21 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds08_1
}
\description{
Dataset for Table 08.1
}
\examples{
ds08_1
}
\keyword{datasets}

