% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds12_1}
\alias{ds12_1}
\title{Dataset for Section 12.1}
\format{A \code{data.frame} with 20 rows and 5 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
\item{y4}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds12_1
}
\description{
Dataset for Section 12.1
}
\examples{
ds12_1
}
\keyword{datasets}

