% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds13_2}
\alias{ds13_2}
\title{13-2 data}
\format{A data.frame with 23 rows and 4 variables:
\describe{
\item{id}{}
\item{p}{}
\item{f}{}
\item{Y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds13_2
}
\description{
ds13_2 data
}
\examples{
ds13_2
}
\keyword{datasets}

