% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds14_1}
\alias{ds14_1}
\title{14-1 data}
\format{A data.frame with 13 rows and 5 variables:
\describe{
\item{id}{}
\item{x}{}
\item{z1}{}
\item{z2}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds14_1
}
\description{
ds14_1 data
}
\examples{
ds14_1
}
\keyword{datasets}

