% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds14_5}
\alias{ds14_5}
\title{Dataset for Section 14.5}
\format{A data.frame with 9 rows and 4 variables:
\describe{
\item{x0}{}
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds14_5
}
\description{
Dataset for Section 14.5
}
\examples{
ds14_5
}
\keyword{datasets}

