% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds15a}
\alias{ds15a}
\title{Dataset for Appendix A, Chapter 15}
\format{A data.frame with 13 rows and 5 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds15a
}
\description{
Dataset for Appendix A, Chapter 15
}
\examples{
ds15a
}
\keyword{datasets}

