% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds18_1}
\alias{ds18_1}
\title{18-1 data}
\format{A data.frame with 10 rows and 4 variables:
\describe{
\item{id}{}
\item{areaa}{}
\item{areab}{}
\item{month}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds18_1
}
\description{
ds18_1 data
}
\examples{
ds18_1
}
\keyword{datasets}

