% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds19_2}
\alias{ds19_2}
\title{19-2 data}
\format{A data.frame with 36 rows and 9 variables:
\describe{
\item{id}{}
\item{z2}{}
\item{z3}{}
\item{z4}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds19_2
}
\description{
ds19_2 data
}
\examples{
ds19_2
}
\keyword{datasets}

