% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds21_7}
\alias{ds21_7}
\title{Dataset for Table 21.7}
\format{A data.frame with 25 rows and 3 variables:
\describe{
\item{y}{}
\item{x6}{}
\item{x8}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds21_7
}
\description{
Dataset for Table 21.7
}
\examples{
ds21_7
}
\keyword{datasets}

