% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds22_2}
\alias{ds22_2}
\title{Dataset for Table 22.2}
\format{A data.frame with 75 rows and 8 variables:
\describe{
\item{n}{}
\item{p0}{}
\item{p1}{}
\item{p2}{}
\item{p3}{}
\item{p4}{}
\item{p5}{}
\item{p6}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds22_2
}
\description{
Dataset for Table 22.2
}
\examples{
ds22_2
}
\keyword{datasets}

