% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds23_10}
\alias{ds23_10}
\title{23-10 data}
\format{A data.frame with 24 rows and 3 variables:
\describe{
\item{rown}{}
\item{column}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds23_10
}
\description{
23-10 data
}
\examples{
ds23_10
}
\keyword{datasets}

