% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds23_2}
\alias{ds23_2}
\title{23-2 data}
\format{A data.frame with 30 rows and 3 variables:
\describe{
\item{id}{}
\item{group}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds23_2
}
\description{
ds23_2 data
}
\examples{
ds23_2
}
\keyword{datasets}

