% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds23_3_3}
\alias{ds23_3_3}
\title{23-3-3 data}
\format{A data.frame with 30 rows and 6 variables:
\describe{
\item{id}{}
\item{x0}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds23_3_3
}
\description{
ds23_3_3 data
}
\examples{
ds23_3_3
}
\keyword{datasets}

