% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds23_7_5}
\alias{ds23_7_5}
\title{23-7-5 data}
\format{A data.frame with 24 rows and 13 variables:
\describe{
\item{id}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{x6}{}
\item{x7}{}
\item{x8}{}
\item{x9}{}
\item{x10}{}
\item{x11}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds23_7_5
}
\description{
ds23_7_5 data
}
\examples{
ds23_7_5
}
\keyword{datasets}

