% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds24_2}
\alias{ds24_2}
\title{Dataset for Table 24.2}
\format{A data.frame with 44 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds24_2
}
\description{
Dataset for Table 24.2
}
\examples{
ds24_2
}
\keyword{datasets}

