% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds25_2}
\alias{ds25_2}
\title{Dataset for Table 25.2}
\format{A data.frame with 20 rows and 3 variables:
\describe{
\item{type}{}
\item{y1974}{}
\item{y1992}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds25_2
}
\description{
Dataset for Table 25.2
}
\examples{
ds25_2
}
\keyword{datasets}

