% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds25_3}
\alias{ds25_3}
\title{25-3 data}
\format{A data.frame with 25 rows and 3 variables:
\describe{
\item{id}{}
\item{diameter}{}
\item{height}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds25_3
}
\description{
ds25_3 data
}
\examples{
ds25_3
}
\keyword{datasets}

