% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dsa01a.R
\docType{data}
\name{dsa01a}
\alias{dsa01a}
\title{Dataset for Appendix A, Chapter 01}
\format{A \code{data.frame} with 25 rows and 10 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{x6}{}
\item{x7}{}
\item{x8}{}
\item{x9}{}
\item{x10}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dsa01a
}
\description{
Dataset for Appendix A, Chapter 01
}
\examples{
dsa01a
}
\keyword{datasets}

