% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa06b.R
\docType{data}
\name{dsa06b}
\alias{dsa06b}
\title{Dataset for Appendix B, Chapter 06}
\format{
A \code{data.frame} with 70 rows and 4 variables:
\describe{
\item{set}{}
\item{x1}{}
\item{x2}{}
\item{y}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dsa06b
}
\description{
Dataset for Appendix B, Chapter 06
}
\examples{
dsa06b
}
\keyword{datasets}
