% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa15a.R
\docType{data}
\name{dsa15a}
\alias{dsa15a}
\title{Dataset for Appendix A, Chapter 15}
\format{
A \code{data.frame} with 13 rows and 5 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{y}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dsa15a
}
\description{
Dataset for Appendix A, Chapter 15
}
\examples{
dsa15a
}
\keyword{datasets}
