% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03a}
\alias{dse03a}
\title{E03A data}
\format{A data.frame with 11 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03a
}
\description{
Data for Exercise A, Chapter 03
}
\examples{
dse03a
}
\keyword{datasets}

