% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03aa}
\alias{dse03aa}
\title{E03AA data}
\format{A data.frame with 7 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03aa
}
\description{
Data for Exercise AA, Chapter 03
}
\examples{
dse03aa
}
\keyword{datasets}

