% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03cc}
\alias{dse03cc}
\title{Dataset for Exercise CC, Chapter 03}
\format{A \code{data.frame} with 10 rows and 10 variables:
\describe{
\item{note}{}
\item{Data}{}
\item{for.}{}
\item{exercise}{}
\item{X3cc}{}
\item{in.}{}
\item{order}{}
\item{X}{}
\item{and}{}
\item{Y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03cc
}
\description{
Dataset for Exercise CC, Chapter 03
}
\examples{
dse03cc
}
\keyword{datasets}

