% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03dd}
\alias{dse03dd}
\title{E03DD data}
\format{A data.frame with 9 rows and 3 variables:
\describe{
\item{id}{}
\item{t}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03dd
}
\description{
dse03dd data
}
\examples{
dse03dd
}
\keyword{datasets}

