% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03e}
\alias{dse03e}
\title{E03E data}
\format{A data.frame with 7 rows and 4 variables:
\describe{
\item{id}{}
\item{x}{}
\item{y1}{}
\item{y2}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03e
}
\description{
dse03e data
}
\examples{
dse03e
}
\keyword{datasets}

