% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03gg}
\alias{dse03gg}
\title{E03GG data}
\format{A data.frame with 5 rows and 3 variables:
\describe{
\item{id}{}
\item{wage}{}
\item{turnover}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03gg
}
\description{
dse03gg data
}
\examples{
dse03gg
}
\keyword{datasets}

