% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03h}
\alias{dse03h}
\title{E03H data}
\format{A data.frame with 14 rows and 3 variables:
\describe{
\item{id}{}
\item{x}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03h
}
\description{
dse03h data
}
\examples{
dse03h
}
\keyword{datasets}

