% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03ii}
\alias{dse03ii}
\title{E03II data}
\format{A data.frame with 32 rows and 3 variables:
\describe{
\item{id}{}
\item{y}{}
\item{x}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03ii
}
\description{
dse03ii data
}
\examples{
dse03ii
}
\keyword{datasets}

