% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dse03j.R
\docType{data}
\name{dse03j}
\alias{dse03j}
\title{Dataset for Exercise J, Chapter 03}
\format{A \code{data.frame} with 15 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03j
}
\description{
Dataset for Exercise J, Chapter 03
}
\examples{
dse03j
}
\keyword{datasets}

