% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03jj}
\alias{dse03jj}
\title{E03JJ data}
\format{A data.frame with 116 rows and 3 variables:
\describe{
\item{id}{}
\item{y}{}
\item{x}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03jj
}
\description{
dse03jj data
}
\examples{
dse03jj
}
\keyword{datasets}

