% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03k}
\alias{dse03k}
\title{E03K data}
\format{A data.frame with 34 rows and 3 variables:
\describe{
\item{id}{}
\item{y}{}
\item{x}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03k
}
\description{
dse03k data
}
\examples{
dse03k
}
\keyword{datasets}

