% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03kk}
\alias{dse03kk}
\title{Dataset for Exercise KK, Chapter 03}
\format{A \code{data.frame} with 12 rows and 2 variables:
\describe{
\item{y1994}{}
\item{y1995}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03kk
}
\description{
Dataset for Exercise KK, Chapter 03
}
\examples{
dse03kk
}
\keyword{datasets}

