% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03r}
\alias{dse03r}
\title{E03R data}
\format{A data.frame with 43 rows and 3 variables:
\describe{
\item{id}{}
\item{t}{}
\item{a}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03r
}
\description{
dse03r data
}
\examples{
dse03r
}
\keyword{datasets}

