% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse03v.R
\docType{data}
\name{dse03v}
\alias{dse03v}
\title{Dataset for Exercise V, Chapter 03}
\format{A \code{data.frame} with 50 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03v
}
\description{
Dataset for Exercise V, Chapter 03
}
\examples{
dse03v
}
\keyword{datasets}

