% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03w}
\alias{dse03w}
\title{E03W data}
\format{A data.frame with 14 rows and 3 variables:
\describe{
\item{id}{}
\item{year}{}
\item{price}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03w
}
\description{
dse03w data
}
\examples{
dse03w
}
\keyword{datasets}

