% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse03x.R
\docType{data}
\name{dse03x}
\alias{dse03x}
\title{Dataset for Exercise X, Chapter 03}
\format{A \code{data.frame} with 11 rows and 6 variables:
\describe{
\item{x123}{}
\item{y1}{}
\item{y2}{}
\item{y3}{}
\item{x4}{}
\item{y4}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03x
}
\description{
Dataset for Exercise X, Chapter 03
}
\examples{
dse03x
}
\keyword{datasets}

