% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03z}
\alias{dse03z}
\title{E03Z data}
\format{A data.frame with 17 rows and 3 variables:
\describe{
\item{id}{}
\item{consumption}{}
\item{population}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03z
}
\description{
dse03z data
}
\examples{
dse03z
}
\keyword{datasets}

