% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse04d}
\alias{dse04d}
\title{E04D data}
\format{A data.frame with 14 rows and 3 variables:
\describe{
\item{id}{}
\item{x}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse04d
}
\description{
dse04d data
}
\examples{
dse04d
}
\keyword{datasets}

