% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse04e}
\alias{dse04e}
\title{E04E data}
\format{A data.frame with 10 rows and 3 variables:
\describe{
\item{id}{}
\item{x}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse04e
}
\description{
dse04e data
}
\examples{
dse04e
}
\keyword{datasets}

