% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse04f}
\alias{dse04f}
\title{E04F data}
\format{A data.frame with 10 rows and 4 variables:
\describe{
\item{id}{}
\item{x}{}
\item{y}{}
\item{xy}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse04f
}
\description{
dse04f data
}
\examples{
dse04f
}
\keyword{datasets}

