% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse06a}
\alias{dse06a}
\title{E06A data}
\format{A data.frame with 11 rows and 5 variables:
\describe{
\item{id}{}
\item{x0}{}
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06a
}
\description{
dse06a data
}
\examples{
dse06a
}
\keyword{datasets}

