% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse06b}
\alias{dse06b}
\title{E06B data}
\format{A data.frame with 12 rows and 4 variables:
\describe{
\item{id}{}
\item{x}{}
\item{y}{}
\item{z}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06b
}
\description{
dse06b data
}
\examples{
dse06b
}
\keyword{datasets}

