% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse06c}
\alias{dse06c}
\title{E06C data}
\format{A data.frame with 15 rows and 4 variables:
\describe{
\item{id}{}
\item{y}{}
\item{x1}{}
\item{x2}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06c
}
\description{
dse06c data
}
\examples{
dse06c
}
\keyword{datasets}

