% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse06d}
\alias{dse06d}
\title{E06D data}
\format{A data.frame with 8 rows and 4 variables:
\describe{
\item{id}{}
\item{y}{}
\item{x1}{}
\item{x2}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06d
}
\description{
dse06d data
}
\examples{
dse06d
}
\keyword{datasets}

