% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse06f}
\alias{dse06f}
\title{Dataset for Exercise F, Chapter 06}
\format{A data.frame with 17 rows and 3 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06f
}
\description{
Dataset for Exercise F, Chapter 06
}
\examples{
dse06f
}
\keyword{datasets}

