% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse06g}
\alias{dse06g}
\title{E06G data}
\format{A data.frame with 17 rows and 4 variables:
\describe{
\item{id}{}
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06g
}
\description{
dse06g data
}
\examples{
dse06g
}
\keyword{datasets}

