% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse06h}
\alias{dse06h}
\title{E06H data}
\format{A data.frame with 13 rows and 4 variables:
\describe{
\item{id}{}
\item{y}{}
\item{x}{}
\item{z}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06h
}
\description{
dse06h data
}
\examples{
dse06h
}
\keyword{datasets}

