% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse06i}
\alias{dse06i}
\title{Dataset for Exercise I, Chapter 06}
\format{A \code{data.frame} with 7 rows and 3 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06i
}
\description{
Dataset for Exercise I, Chapter 06
}
\examples{
dse06i
}
\keyword{datasets}

