% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse06l}
\alias{dse06l}
\title{E06L data}
\format{A data.frame with 9 rows and 5 variables:
\describe{
\item{id}{}
\item{batchn}{}
\item{t}{}
\item{c}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06l
}
\description{
dse06l data
}
\examples{
dse06l
}
\keyword{datasets}

