% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse06z}
\alias{dse06z}
\title{E06Z data}
\format{A data.frame with 19 rows and 3 variables:
\describe{
\item{id}{}
\item{x}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06z
}
\description{
dse06z data
}
\examples{
dse06z
}
\keyword{datasets}

