% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse07b}
\alias{dse07b}
\title{E07B data}
\format{A data.frame with 48 rows and 4 variables:
\describe{
\item{id}{}
\item{set1}{}
\item{set2}{}
\item{set3}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse07b
}
\description{
dse07b data
}
\examples{
dse07b
}
\keyword{datasets}

