% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse08b}
\alias{dse08b}
\title{E08B data}
\format{A data.frame with 5 rows and 3 variables:
\describe{
\item{id}{}
\item{x}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse08b
}
\description{
dse08b data
}
\examples{
dse08b
}
\keyword{datasets}

