% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse12a}
\alias{dse12a}
\title{E12A data}
\format{A data.frame with 18 rows and 6 variables:
\describe{
\item{id}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse12a
}
\description{
dse12a data
}
\examples{
dse12a
}
\keyword{datasets}

