% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse12c}
\alias{dse12c}
\title{E12C data}
\format{A data.frame with 24 rows and 3 variables:
\describe{
\item{id}{}
\item{t}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse12c
}
\description{
dse12c data
}
\examples{
dse12c
}
\keyword{datasets}

