% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse12d}
\alias{dse12d}
\title{Dataset for Exercise D, Chapter 12}
\format{A \code{data.frame} with 15 rows and 6 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y1}{}
\item{y2}{}
\item{y3}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse12d
}
\description{
Dataset for Exercise D, Chapter 12
}
\examples{
dse12d
}
\keyword{datasets}

