% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse12e}
\alias{dse12e}
\title{E12E data}
\format{A data.frame with 28 rows and 6 variables:
\describe{
\item{id}{}
\item{fat}{}
\item{flour}{}
\item{water}{}
\item{rpm}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse12e
}
\description{
dse12e data
}
\examples{
dse12e
}
\keyword{datasets}

