% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse12h}
\alias{dse12h}
\title{E12H data}
\format{A data.frame with 27 rows and 5 variables:
\describe{
\item{id}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse12h
}
\description{
dse12h data
}
\examples{
dse12h
}
\keyword{datasets}

