% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse13a}
\alias{dse13a}
\title{Dataset for Exercise A, Chapter 13}
\format{A \code{data.frame} with 6 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse13a
}
\description{
Dataset for Exercise A, Chapter 13
}
\examples{
dse13a
}
\keyword{datasets}

