% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse13b.R
\docType{data}
\name{dse13b}
\alias{dse13b}
\title{Dataset for Exercise B, Chapter 13}
\format{
A \code{data.frame} with 47 rows and 4 variables:
\describe{
\item{filler}{}
\item{naphthenic_oil}{}
\item{y}{}
\item{ytype}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse13b
}
\description{
Dataset for Exercise B, Chapter 13
}
\examples{
dse13b
}
\keyword{datasets}
