% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse13c}
\alias{dse13c}
\title{E13C data}
\format{A data.frame with 11 rows and 7 variables:
\describe{
\item{id}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{y1}{}
\item{y2}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse13c
}
\description{
dse13c data
}
\examples{
dse13c
}
\keyword{datasets}

