% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse13d}
\alias{dse13d}
\title{E13D data}
\format{A data.frame with 24 rows and 5 variables:
\describe{
\item{id}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse13d
}
\description{
dse13d data
}
\examples{
dse13d
}
\keyword{datasets}

