% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse13e}
\alias{dse13e}
\title{E13E data}
\format{A data.frame with 43 rows and 3 variables:
\describe{
\item{id}{}
\item{t}{}
\item{m}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse13e
}
\description{
dse13e data
}
\examples{
dse13e
}
\keyword{datasets}

