% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse13g}
\alias{dse13g}
\title{E13G data}
\format{A data.frame with 35 rows and 5 variables:
\describe{
\item{id}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse13g
}
\description{
dse13g data
}
\examples{
dse13g
}
\keyword{datasets}

