% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse13h}
\alias{dse13h}
\title{E13H data}
\format{A data.frame with 8 rows and 6 variables:
\describe{
\item{id}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse13h
}
\description{
dse13h data
}
\examples{
dse13h
}
\keyword{datasets}

