% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse14b}
\alias{dse14b}
\title{Dataset for Exercise B, Chapter 14}
\format{A \code{data.frame} with 18 rows and 6 variables:
\describe{
\item{y}{}
\item{z1}{}
\item{z2}{}
\item{z3}{}
\item{nrate}{}
\item{x}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse14b
}
\description{
Dataset for Exercise B, Chapter 14
}
\examples{
dse14b
}
\keyword{datasets}

