% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse14c.R
\docType{data}
\name{dse14c}
\alias{dse14c}
\title{Dataset for Exercise C, Chapter 14}
\format{A \code{data.frame} with 9 rows and 4 variables:
\describe{
\item{row}{}
\item{y}{}
\item{operator}{}
\item{speed}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse14c
}
\description{
Dataset for Exercise C, Chapter 14
}
\examples{
dse14c
}
\keyword{datasets}

