% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse14d}
\alias{dse14d}
\title{Dataset for Exercise D, Chapter 14}
\format{A \code{data.frame} with 6 rows and 6 variables:
\describe{
\item{z0}{}
\item{z1}{}
\item{z2}{}
\item{z3}{}
\item{z4}{}
\item{z5}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse14d
}
\description{
Dataset for Exercise D, Chapter 14
}
\examples{
dse14d
}
\keyword{datasets}

