% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse14f.R
\docType{data}
\name{dse14f}
\alias{dse14f}
\title{Dataset for Exercise F, Chapter 14}
\format{
A \code{data.frame} with 6 rows and 6 variables:
\describe{
\item{z0}{}
\item{z1}{}
\item{z2}{}
\item{z3}{}
\item{z4}{}
\item{z5}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse14f
}
\description{
Dataset for Exercise F, Chapter 14
}
\examples{
dse14f
}
\keyword{datasets}
