% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse14g}
\alias{dse14g}
\title{Dataset for Exercise G, Chapter 14}
\format{A \code{data.frame} with 20 rows and 5 variables:
\describe{
\item{system}{}
\item{x0}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse14g
}
\description{
Dataset for Exercise G, Chapter 14
}
\examples{
dse14g
}
\keyword{datasets}

