% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse14j}
\alias{dse14j}
\title{Dataset for Exercise J, Chapter 14}
\format{A data.frame with 72 rows and 2 variables:
\describe{
\item{wh}{}
\item{age}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse14j
}
\description{
Dataset for Exercise J, Chapter 14
}
\examples{
dse14j
}
\keyword{datasets}

