% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse14q.R
\docType{data}
\name{dse14q}
\alias{dse14q}
\title{Dataset for Exercise Q, Chapter 14}
\format{
A \code{data.frame} with 8 rows and 6 variables:
\describe{
\item{x1}{}
\item{y1}{}
\item{x2}{}
\item{y2}{}
\item{x3}{}
\item{y3}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse14q
}
\description{
Dataset for Exercise Q, Chapter 14
}
\examples{
dse14q
}
\keyword{datasets}
