% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse14r.R
\docType{data}
\name{dse14r}
\alias{dse14r}
\title{Dataset for Exercise R, Chapter 14}
\format{A \code{data.frame} with 48 rows and 4 variables:
\describe{
\item{boot}{}
\item{subject}{}
\item{temperature}{}
\item{drop}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse14r
}
\description{
Dataset for Exercise R, Chapter 14
}
\examples{
dse14r
}
\keyword{datasets}

