% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse14s}
\alias{dse14s}
\title{E14S data}
\format{A data.frame with 8 rows and 5 variables:
\describe{
\item{id}{}
\item{x}{}
\item{z}{}
\item{zx}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse14s
}
\description{
dse14s data
}
\examples{
dse14s
}
\keyword{datasets}

