% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse14t}
\alias{dse14t}
\title{E14T data}
\format{A data.frame with 15 rows and 5 variables:
\describe{
\item{id}{}
\item{x}{}
\item{z}{}
\item{xz}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse14t
}
\description{
dse14t data
}
\examples{
dse14t
}
\keyword{datasets}

