% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse14u}
\alias{dse14u}
\title{E14U data}
\format{A data.frame with 15 rows and 5 variables:
\describe{
\item{id}{}
\item{z1}{}
\item{z2}{}
\item{z3}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse14u
}
\description{
dse14u data
}
\examples{
dse14u
}
\keyword{datasets}

