% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse15e}
\alias{dse15e}
\title{E15E data}
\format{A data.frame with 90 rows and 9 variables:
\describe{
\item{id}{}
\item{y}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{x6}{}
\item{x7}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse15e
}
\description{
dse15e data
}
\examples{
dse15e
}
\keyword{datasets}

