% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dse15h.R
\docType{data}
\name{dse15h}
\alias{dse15h}
\title{Dataset for Exercise H, Chapter 15}
\format{A \code{data.frame} with 33 rows and 10 variables:
\describe{
\item{year}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{x6}{}
\item{x7}{}
\item{x8}{}
\item{x9}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse15h
}
\description{
Dataset for Exercise H, Chapter 15
}
\examples{
dse15h
}
\keyword{datasets}

