% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse15k}
\alias{dse15k}
\title{E15K data}
\format{A data.frame with 16 rows and 6 variables:
\describe{
\item{y1}{}
\item{y2}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse15k
}
\description{
Data for Exercise K, Chapter 15
}
\examples{
dse15k
}
\keyword{datasets}

