% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse15l.R
\docType{data}
\name{dse15l}
\alias{dse15l}
\title{Dataset for Exercise L, Chapter 15}
\format{A \code{data.frame} with 15 rows and 8 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{x6}{}
\item{x7}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse15l
}
\description{
Dataset for Exercise L, Chapter 15
}
\examples{
dse15l
}
\keyword{datasets}

