% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse15n}
\alias{dse15n}
\title{E15N data}
\format{A data.frame with 21 rows and 5 variables:
\describe{
\item{id}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse15n
}
\description{
dse15n data
}
\examples{
dse15n
}
\keyword{datasets}

