% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse15p}
\alias{dse15p}
\title{E15P data}
\format{A data.frame with 16 rows and 7 variables:
\describe{
\item{id}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse15p
}
\description{
dse15p data
}
\examples{
dse15p
}
\keyword{datasets}

