% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse16b}
\alias{dse16b}
\title{Dataset for Exercise B, Chapter 16}
\format{A \code{data.frame} with 6 rows and 4 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse16b
}
\description{
Dataset for Exercise B, Chapter 16
}
\examples{
dse16b
}
\keyword{datasets}

