% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse16c}
\alias{dse16c}
\title{E16C data}
\format{A data.frame with 7 rows and 4 variables:
\describe{
\item{id}{}
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse16c
}
\description{
dse16c data
}
\examples{
dse16c
}
\keyword{datasets}

