% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse17b}
\alias{dse17b}
\title{E17B data}
\format{A data.frame with 8 rows and 3 variables:
\describe{
\item{id}{}
\item{x}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse17b
}
\description{
dse17b data
}
\examples{
dse17b
}
\keyword{datasets}

