% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse22a.R
\docType{data}
\name{dse22a}
\alias{dse22a}
\title{Dataset for Exercise A, Chapter 22}
\format{
A \code{data.frame} with 11 rows and 9 variables:
\describe{
\item{a}{}
\item{b}{}
\item{p0}{}
\item{p1}{}
\item{p2}{}
\item{p3}{}
\item{p4}{}
\item{p5}{}
\item{p6}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse22a
}
\description{
Dataset for Exercise A, Chapter 22
}
\examples{
dse22a
}
\keyword{datasets}
