% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse22b.R
\docType{data}
\name{dse22b}
\alias{dse22b}
\title{Dataset for Exercise B, Chapter 22}
\format{A \code{data.frame} with 6 rows and 4 variables:
\describe{
\item{y}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse22b
}
\description{
Dataset for Exercise B, Chapter 22
}
\examples{
dse22b
}
\keyword{datasets}

