% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dse22c.R
\docType{data}
\name{dse22c}
\alias{dse22c}
\title{Dataset for Exercise C, Chapter 22}
\format{A \code{data.frame} with 20 rows and 7 variables:
\describe{
\item{p1}{}
\item{p2}{}
\item{p3}{}
\item{p4}{}
\item{p5}{}
\item{p6}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse22c
}
\description{
Dataset for Exercise C, Chapter 22
}
\examples{
dse22c
}
\keyword{datasets}

