% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse22d.R
\docType{data}
\name{dse22d}
\alias{dse22d}
\title{Dataset for Exercise D, Chapter 22}
\format{
A \code{data.frame} with 10 rows and 9 variables:
\describe{
\item{time}{}
\item{weight_diff}{}
\item{p0}{}
\item{p1}{}
\item{p2}{}
\item{p3}{}
\item{p4}{}
\item{p5}{}
\item{p6}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse22d
}
\description{
Dataset for Exercise D, Chapter 22
}
\examples{
dse22d
}
\keyword{datasets}
