% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse22f}
\alias{dse22f}
\title{E22F data}
\format{A data.frame with 12 rows and 8 variables:
\describe{
\item{id}{}
\item{y}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{x6}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse22f
}
\description{
dse22f data
}
\examples{
dse22f
}
\keyword{datasets}

