% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse22g}
\alias{dse22g}
\title{E22G data}
\format{A data.frame with 24 rows and 7 variables:
\describe{
\item{id}{}
\item{age}{}
\item{n}{}
\item{weight}{}
\item{sd}{}
\item{z1}{}
\item{z2}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse22g
}
\description{
dse22g data
}
\examples{
dse22g
}
\keyword{datasets}

