% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse23a_a.R
\docType{data}
\name{dse23a_a}
\alias{dse23a_a}
\title{Dataset for Exercise A-a, Chapter 23}
\format{A \code{data.frame} with 18 rows and 7 variables:
\describe{
\item{pressure}{}
\item{time}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse23a_a
}
\description{
Dataset for Exercise A-a, Chapter 23
}
\examples{
dse23a_a
}
\keyword{datasets}

