% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dse23d.R
\docType{data}
\name{dse23d}
\alias{dse23d}
\title{Dataset for Exercise D, Chapter 23}
\format{A \code{data.frame} with 19 rows and 2 variables:
\describe{
\item{pcal}{}
\item{bstrength}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse23d
}
\description{
Dataset for Exercise D, Chapter 23
}
\examples{
dse23d
}
\keyword{datasets}

