% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse23e}
\alias{dse23e}
\title{Dataset for Exercise E, Chapter 23}
\format{A \code{data.frame} with 16 rows and 9 variables:
\describe{
\item{al}{}
\item{bl}{}
\item{cl}{}
\item{dl}{}
\item{a}{}
\item{b}{}
\item{c}{}
\item{d}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse23e
}
\description{
Dataset for Exercise E, Chapter 23
}
\examples{
dse23e
}
\keyword{datasets}

