% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse23f.R
\docType{data}
\name{dse23f}
\alias{dse23f}
\title{Dataset for Exercise F, Chapter 23}
\format{A \code{data.frame} with 9 rows and 4 variables:
\describe{
\item{group}{}
\item{z1}{}
\item{z2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse23f
}
\description{
Dataset for Exercise F, Chapter 23
}
\examples{
dse23f
}
\keyword{datasets}

