% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse23g_1.R
\docType{data}
\name{dse23g_1}
\alias{dse23g_1}
\title{Dataset for Exercise G-1, Chapter 23}
\format{A \code{data.frame} with 15 rows and 9 variables:
\describe{
\item{level}{}
\item{tube}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{x6}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse23g_1
}
\description{
Dataset for Exercise G-1, Chapter 23
}
\examples{
dse23g_1
}
\keyword{datasets}

