% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse23h.R
\docType{data}
\name{dse23h}
\alias{dse23h}
\title{Dataset for Exercise H, Chapter 23}
\format{
A \code{data.frame} with 16 rows and 6 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x2x3}{}
\item{matings}{}
\item{y}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse23h
}
\description{
Dataset for Exercise H, Chapter 23
}
\examples{
dse23h
}
\keyword{datasets}
