% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24a}
\alias{dse24a}
\title{E24A data}
\format{A data.frame with 3 rows and 3 variables:
\describe{
\item{id}{}
\item{t}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24a
}
\description{
dse24a data
}
\examples{
dse24a
}
\keyword{datasets}

