% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24b}
\alias{dse24b}
\title{E24B data}
\format{A data.frame with 12 rows and 2 variables:
\describe{
\item{t}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24b
}
\description{
Data for Exercise B, Chapter 24
}
\examples{
dse24b
}
\keyword{datasets}

