% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24c}
\alias{dse24c}
\title{E24C data}
\format{A data.frame with 4 rows and 3 variables:
\describe{
\item{id}{}
\item{x}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24c
}
\description{
dse24c data
}
\examples{
dse24c
}
\keyword{datasets}

