% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24d}
\alias{dse24d}
\title{E24D data}
\format{A data.frame with 5 rows and 3 variables:
\describe{
\item{id}{}
\item{x}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24d
}
\description{
dse24d data
}
\examples{
dse24d
}
\keyword{datasets}

