% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24e}
\alias{dse24e}
\title{E24E data}
\format{A data.frame with 14 rows and 2 variables:
\describe{
\item{t}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24e
}
\description{
Data for Exercise E, Chapter 24
}
\examples{
dse24e
}
\keyword{datasets}

