% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24i}
\alias{dse24i}
\title{E24I data}
\format{A data.frame with 8 rows and 4 variables:
\describe{
\item{id}{}
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24i
}
\description{
dse24i data
}
\examples{
dse24i
}
\keyword{datasets}

