% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24j}
\alias{dse24j}
\title{E24J data}
\format{A data.frame with 13 rows and 4 variables:
\describe{
\item{id}{}
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24j
}
\description{
dse24j data
}
\examples{
dse24j
}
\keyword{datasets}

