% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24l}
\alias{dse24l}
\title{Dataset for Exercise L, Chapter 24}
\format{A data.frame with 55 rows and 3 variables:
\describe{
\item{t}{}
\item{y}{}
\item{set}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24l
}
\description{
Dataset for Exercise L, Chapter 24
}
\examples{
dse24l
}
\keyword{datasets}

