% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24o}
\alias{dse24o}
\title{E24O data}
\format{A data.frame with 4 rows and 3 variables:
\describe{
\item{id}{}
\item{x}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24o
}
\description{
dse24o data
}
\examples{
dse24o
}
\keyword{datasets}

