% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dse24p.R
\docType{data}
\name{dse24p}
\alias{dse24p}
\title{Dataset for Exercise P, Chapter 24}
\format{A \code{data.frame} with 5 rows and 3 variables:
\describe{
\item{x}{}
\item{y}{}
\item{z}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24p
}
\description{
Dataset for Exercise P, Chapter 24
}
\examples{
dse24p
}
\keyword{datasets}

