% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dsq23_3_3.R
\docType{data}
\name{dsq23_3_3}
\alias{dsq23_3_3}
\title{Dataset for Equation 23.3.3}
\format{A \code{data.frame} with 30 rows and 5 variables:
\describe{
\item{x0}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dsq23_3_3
}
\description{
Dataset for Equation 23.3.3
}
\examples{
dsq23_3_3
}
\keyword{datasets}

