% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dsq23_3_7}
\alias{dsq23_3_7}
\title{Dataset for Equation 23.3.7}
\format{A \code{data.frame} with 30 rows and 4 variables:
\describe{
\item{x0}{}
\item{z1}{}
\item{z2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dsq23_3_7
}
\description{
Dataset for Equation 23.3.7
}
\examples{
dsq23_3_7
}
\keyword{datasets}

