% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dss217}
\alias{dss217}
\title{Dataset for Section 21.7}
\format{A \code{data.frame} with 25 rows and 3 variables:
\describe{
\item{y}{}
\item{x6}{}
\item{x8}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dss217
}
\description{
Dataset for Section 21.7
}
\examples{
dss217
}
\keyword{datasets}

