% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dss222.R
\docType{data}
\name{dss222}
\alias{dss222}
\title{Dataset for Section 22.2}
\format{A \code{data.frame} with 75 rows and 8 variables:
\describe{
\item{n}{}
\item{p0}{}
\item{p1}{}
\item{p2}{}
\item{p3}{}
\item{p4}{}
\item{p5}{}
\item{p6}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dss222
}
\description{
Dataset for Section 22.2
}
\examples{
dss222
}
\keyword{datasets}

