% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dss2310}
\alias{dss2310}
\title{Dataset for Section 23.10}
\format{A \code{data.frame} with 24 rows and 3 variables:
\describe{
\item{rown}{}
\item{column}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dss2310
}
\description{
Dataset for Section 23.10
}
\examples{
dss2310
}
\keyword{datasets}

