% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst021.R
\docType{data}
\name{dst021}
\alias{dst021}
\title{Dataset from Table 02.1}
\format{
A \code{data.frame} with 23 rows and 3 variables:
\describe{
\item{time_order}{}
\item{y}{}
\item{x}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst021
}
\description{
Dataset from Table 02.1
}
\examples{
dst021
}
\keyword{datasets}
