% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dst02_1}
\alias{dst02_1}
\title{Dataset from Table 02.1}
\format{A \code{data.frame} with 23 rows and 3 variables:
\describe{
\item{time_order}{}
\item{y}{}
\item{x}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst02_1
}
\description{
Dataset from Table 02.1
}
\examples{
dst02_1
}
\keyword{datasets}

