% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst032.R
\docType{data}
\name{dst032}
\alias{dst032}
\title{Dataset from Table 03.2}
\format{
A \code{data.frame} with 49 rows and 3 variables:
\describe{
\item{v26}{}
\item{logs}{}
\item{z}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst032
}
\description{
Dataset from Table 03.2
}
\examples{
dst032
}
\keyword{datasets}
