% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dst033.R
\docType{data}
\name{dst033}
\alias{dst033}
\title{Dataset from Table 03.3}
\format{A \code{data.frame} with 20 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst033
}
\description{
Dataset from Table 03.3
}
\examples{
dst033
}
\keyword{datasets}

