% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dst03_3}
\alias{dst03_3}
\title{Dataset for Table 03.3}
\format{A \code{data.frame} with 20 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst03_3
}
\description{
Dataset for Table 03.3
}
\examples{
dst03_3
}
\keyword{datasets}

