% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dst081}
\alias{dst081}
\title{Dataset from Table 08.1}
\format{A \code{data.frame} with 21 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst081
}
\description{
Dataset from Table 08.1
}
\examples{
dst081
}
\keyword{datasets}

