% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst121.R
\docType{data}
\name{dst121}
\alias{dst121}
\title{Dataset from Table 12.1}
\format{
A \code{data.frame} with 20 rows and 5 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
\item{y4}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst121
}
\description{
Dataset from Table 12.1
}
\examples{
dst121
}
\keyword{datasets}
