% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dst132.R
\docType{data}
\name{dst132}
\alias{dst132}
\title{Dataset from Table 13.2}
\format{A \code{data.frame} with 23 rows and 3 variables:
\describe{
\item{p}{}
\item{f}{}
\item{Y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst132
}
\description{
Dataset from Table 13.2
}
\examples{
dst132
}
\keyword{datasets}

