% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dst134.R
\docType{data}
\name{dst134}
\alias{dst134}
\title{Dataset from Table 13.4}
\format{A \code{data.frame} with 23 rows and 3 variables:
\describe{
\item{p}{}
\item{f}{}
\item{lny}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst134
}
\description{
Dataset from Table 13.4
}
\examples{
dst134
}
\keyword{datasets}

