% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dst13_2}
\alias{dst13_2}
\title{Dataset from Table 13.2}
\format{A \code{data.frame} with 23 rows and 3 variables:
\describe{
\item{p}{}
\item{f}{}
\item{Y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst13_2
}
\description{
Dataset from Table 13.2
}
\examples{
dst13_2
}
\keyword{datasets}

