% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dst141.R
\docType{data}
\name{dst141}
\alias{dst141}
\title{Dataset from Table 14.1}
\format{A \code{data.frame} with 13 rows and 4 variables:
\describe{
\item{x}{}
\item{z1}{}
\item{z2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst141
}
\description{
Dataset from Table 14.1
}
\examples{
dst141
}
\keyword{datasets}

