% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dst144.R
\docType{data}
\name{dst144}
\alias{dst144}
\title{Dataset from Table 14.4}
\format{A \code{data.frame} with 9 rows and 4 variables:
\describe{
\item{x0}{}
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst144
}
\description{
Dataset from Table 14.4
}
\examples{
dst144
}
\keyword{datasets}

