% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dst145}
\alias{dst145}
\title{Dataset from Table 14.5}
\format{A \code{data.frame} with 9 rows and 4 variables:
\describe{
\item{x0}{}
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst145
}
\description{
Dataset from Table 14.5
}
\examples{
dst145
}
\keyword{datasets}

