% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dst146.R
\docType{data}
\name{dst146}
\alias{dst146}
\title{Dataset from Table 14.6}
\format{A \code{data.frame} with 9 rows and 5 variables:
\describe{
\item{x0}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst146
}
\description{
Dataset from Table 14.6
}
\examples{
dst146
}
\keyword{datasets}

