% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dst14_1}
\alias{dst14_1}
\title{Dataset from Table 14.1}
\format{A \code{data.frame} with 13 rows and 4 variables:
\describe{
\item{x}{}
\item{z1}{}
\item{z2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst14_1
}
\description{
Dataset from Table 14.1
}
\examples{
dst14_1
}
\keyword{datasets}

