% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dst14_6}
\alias{dst14_6}
\title{Dataset from Table 14.6}
\format{A \code{data.frame} with 9 rows and 5 variables:
\describe{
\item{x0}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst14_6
}
\description{
Dataset from Table 14.6
}
\examples{
dst14_6
}
\keyword{datasets}

