% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst181.R
\docType{data}
\name{dst181}
\alias{dst181}
\title{Dataset from Table 18.1}
\format{
A \code{data.frame} with 10 rows and 3 variables:
\describe{
\item{areaa}{}
\item{areab}{}
\item{month}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst181
}
\description{
Dataset from Table 18.1
}
\examples{
dst181
}
\keyword{datasets}
