% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dst191}
\alias{dst191}
\title{Dataset from Table 19.1}
\format{A \code{data.frame} with 10 rows and 4 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst191
}
\description{
Dataset from Table 19.1
}
\examples{
dst191
}
\keyword{datasets}

