% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dst19_2}
\alias{dst19_2}
\title{Dataset from Table 19.2}
\format{A \code{data.frame} with 36 rows and 8 variables:
\describe{
\item{z2}{}
\item{z3}{}
\item{z4}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst19_2
}
\description{
Dataset from Table 19.2
}
\examples{
dst19_2
}
\keyword{datasets}

