% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dst232.R
\docType{data}
\name{dst232}
\alias{dst232}
\title{Dataset from Table 23.2}
\format{A \code{data.frame} with 30 rows and 2 variables:
\describe{
\item{group}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst232
}
\description{
Dataset from Table 23.2
}
\examples{
dst232
}
\keyword{datasets}

