% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dst23_2}
\alias{dst23_2}
\title{Dataset from Table 23.2}
\format{A \code{data.frame} with 30 rows and 2 variables:
\describe{
\item{group}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst23_2
}
\description{
Dataset from Table 23.2
}
\examples{
dst23_2
}
\keyword{datasets}

