% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dst24_2}
\alias{dst24_2}
\title{Dataset from Table 24.2}
\format{A \code{data.frame} with 44 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst24_2
}
\description{
Dataset from Table 24.2
}
\examples{
dst24_2
}
\keyword{datasets}

