% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst252.R
\docType{data}
\name{dst252}
\alias{dst252}
\title{Dataset from Table 25.2}
\format{
A \code{data.frame} with 20 rows and 3 variables:
\describe{
\item{type}{}
\item{y1974}{}
\item{y1992}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst252
}
\description{
Dataset from Table 25.2
}
\examples{
dst252
}
\keyword{datasets}
