% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dst253}
\alias{dst253}
\title{Dataset from Table 25.3}
\format{A \code{data.frame} with 25 rows and 2 variables:
\describe{
\item{height}{}
\item{diameter}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst253
}
\description{
Dataset from Table 25.3
}
\examples{
dst253
}
\keyword{datasets}

