% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dst25_2}
\alias{dst25_2}
\title{Dataset from Table 25.2}
\format{A \code{data.frame} with 20 rows and 3 variables:
\describe{
\item{type}{}
\item{y1974}{}
\item{y1992}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst25_2
}
\description{
Dataset from Table 25.2
}
\examples{
dst25_2
}
\keyword{datasets}

