% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst261.R
\docType{data}
\name{dst261}
\alias{dst261}
\title{Dataset from Table 26.1}
\format{
A \code{data.frame} with 10 rows and 2 variables:
\describe{
\item{x1974}{}
\item{y1992}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst261
}
\description{
Dataset from Table 26.1
}
\examples{
dst261
}
\keyword{datasets}
