% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dst26_1}
\alias{dst26_1}
\title{Dataset from Table 26.1}
\format{A \code{data.frame} with 10 rows and 2 variables:
\describe{
\item{x1974}{}
\item{y1992}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dst26_1
}
\description{
Dataset from Table 26.1
}
\examples{
dst26_1
}
\keyword{datasets}

