% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsx161.R
\docType{data}
\name{dsx161}
\alias{dsx161}
\title{Dataset for Example 1, Chapter 16}
\format{
A \code{matrix} with 6 rows and 5 columns.
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dsx161
}
\description{
Dataset for Example 1, Chapter 16
}
\examples{
dsx161
}
\keyword{datasets}
