% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{aprean3}
\alias{aprean3}
\alias{aprean3-package}
\title{Datasets from Draper and Smith, "Applied Regression Analysis" (3rd ed.,
1998)}
\source{
Draper, N.R., Smith, H. (1998) Applied Regression Analysis, 3rd
ed., New York: Wiley
}
\description{
This package is an unofficial companion to the textbook "Applied
Regression Analysis" by N.R. Draper and H. Smith (3rd ed., 1998).
}
\details{
It includes all the published datasets, both for easy reproducibility
and algorithms benchmarking purposes.

Dataset names are lowercased and prefixed by \code{dsX}, where \code{X} is:
\itemize{
\item \code{e} for Exercises
\item \code{x} for eXamples
\item \code{q} for eQuations
\item \code{a} for Appendixes
\item \code{s} for Sections.
}
Variables have lower-case name.
}

