% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds01a}
\alias{ds01a}
\title{01A data}
\format{A data.frame with 25 rows and 11 variables:
\describe{
\item{id}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{x6}{}
\item{x7}{}
\item{x8}{}
\item{x9}{}
\item{x10}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds01a
}
\description{
ds01a data
}
\examples{
ds01a
}
\keyword{datasets}

