% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds13_2}
\alias{ds13_2}
\title{Dataset for Section 13.2}
\format{A data.frame with 23 rows and 3 variables:
\describe{
\item{p}{}
\item{f}{}
\item{Y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds13_2
}
\description{
Dataset for Section 13.2
}
\examples{
ds13_2
}
\keyword{datasets}

