% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds13_4}
\alias{ds13_4}
\title{13-4 data}
\format{A data.frame with 23 rows and 4 variables:
\describe{
\item{id}{}
\item{p}{}
\item{f}{}
\item{lny}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds13_4
}
\description{
ds13_4 data
}
\examples{
ds13_4
}
\keyword{datasets}

