% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds13_4}
\alias{ds13_4}
\title{Dataset for Table 13.4}
\format{A data.frame with 23 rows and 3 variables:
\describe{
\item{p}{}
\item{f}{}
\item{lny}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds13_4
}
\description{
Dataset for Table 13.4
}
\examples{
ds13_4
}
\keyword{datasets}

