% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds14_1}
\alias{ds14_1}
\title{Dataset for Section 14.1}
\format{A data.frame with 13 rows and 4 variables:
\describe{
\item{x}{}
\item{z1}{}
\item{z2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds14_1
}
\description{
Dataset for Section 14.1
}
\examples{
ds14_1
}
\keyword{datasets}

