% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds16ex1}
\alias{ds16ex1}
\title{Dataset for Example 1, Chapter 16}
\format{A \code{matrix} with 6 rows and 5 columns.}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds16ex1
}
\description{
Dataset for Example 1, Chapter 16
}
\examples{
ds16ex1
}
\keyword{datasets}

