% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds18_1}
\alias{ds18_1}
\title{Dataset for Table 18.1}
\format{A data.frame with 10 rows and 3 variables:
\describe{
\item{areaa}{}
\item{areab}{}
\item{month}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds18_1
}
\description{
Dataset for Table 18.1
}
\examples{
ds18_1
}
\keyword{datasets}

