% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds19_1}
\alias{ds19_1}
\title{19-1 data}
\format{A data.frame with 10 rows and 4 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds19_1
}
\description{
19-1 data
}
\examples{
ds19_1
}
\keyword{datasets}

