% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds21_7}
\alias{ds21_7}
\title{21-7 data}
\format{A data.frame with 25 rows and 4 variables:
\describe{
\item{id}{}
\item{y}{}
\item{x6}{}
\item{x8}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds21_7
}
\description{
ds21_7 data
}
\examples{
ds21_7
}
\keyword{datasets}

