% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds23_10}
\alias{ds23_10}
\title{23-10 data}
\format{A data.frame with 24 rows and 4 variables:
\describe{
\item{id}{}
\item{rown}{}
\item{column}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds23_10
}
\description{
ds23_10 data
}
\examples{
ds23_10
}
\keyword{datasets}

