% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds23_2}
\alias{ds23_2}
\title{Dataset for Table 23.2}
\format{A data.frame with 30 rows and 2 variables:
\describe{
\item{group}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds23_2
}
\description{
Dataset for Table 23.2
}
\examples{
ds23_2
}
\keyword{datasets}

