% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds23_7_5}
\alias{ds23_7_5}
\title{Dataset for Equation 23.7.5}
\format{A \code{data.frame} with 24 rows and 12 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{x6}{}
\item{x7}{}
\item{x8}{}
\item{x9}{}
\item{x10}{}
\item{x11}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds23_7_5
}
\description{
Dataset for Equation 23.7.5
}
\examples{
ds23_7_5
}
\keyword{datasets}

