% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds24_2}
\alias{ds24_2}
\title{24-2 data}
\format{A data.frame with 44 rows and 3 variables:
\describe{
\item{id}{}
\item{x}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds24_2
}
\description{
ds24_2 data
}
\examples{
ds24_2
}
\keyword{datasets}

