% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ds25_3}
\alias{ds25_3}
\title{Dataset for Section 25.3}
\format{A data.frame with 25 rows and 2 variables:
\describe{
\item{diameter}{}
\item{height}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
ds25_3
}
\description{
Dataset for Section 25.3
}
\examples{
ds25_3
}
\keyword{datasets}

