% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse03a.R
\docType{data}
\name{dse03a}
\alias{dse03a}
\title{Dataset for Exercise A, Chapter 03}
\format{A \code{data.frame} with 11 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03a
}
\description{
Dataset for Exercise A, Chapter 03
}
\examples{
dse03a
}
\keyword{datasets}

