% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse03bb.R
\docType{data}
\name{dse03bb}
\alias{dse03bb}
\title{Dataset for Exercise BB, Chapter 03}
\format{A \code{data.frame} with 10 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03bb
}
\description{
Dataset for Exercise BB, Chapter 03
}
\examples{
dse03bb
}
\keyword{datasets}

