% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03bb}
\alias{dse03bb}
\title{E03BB data}
\format{A data.frame with 10 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03bb
}
\description{
Data for Exercise BB, Chapter 03
}
\examples{
dse03bb
}
\keyword{datasets}

