% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse03dd.R
\docType{data}
\name{dse03dd}
\alias{dse03dd}
\title{Dataset for Exercise DD, Chapter 03}
\format{A \code{data.frame} with 9 rows and 2 variables:
\describe{
\item{t}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03dd
}
\description{
Dataset for Exercise DD, Chapter 03
}
\examples{
dse03dd
}
\keyword{datasets}

