% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03dd}
\alias{dse03dd}
\title{E03DD data}
\format{A data.frame with 9 rows and 2 variables:
\describe{
\item{t}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03dd
}
\description{
Data for Exercise DD, Chapter 03
}
\examples{
dse03dd
}
\keyword{datasets}

