% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse03e.R
\docType{data}
\name{dse03e}
\alias{dse03e}
\title{Dataset for Exercise E, Chapter 03}
\format{A \code{data.frame} with 7 rows and 3 variables:
\describe{
\item{x}{}
\item{y1}{}
\item{y2}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03e
}
\description{
Dataset for Exercise E, Chapter 03
}
\examples{
dse03e
}
\keyword{datasets}

