% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03e}
\alias{dse03e}
\title{E03E data}
\format{A data.frame with 7 rows and 3 variables:
\describe{
\item{x}{}
\item{y1}{}
\item{y2}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03e
}
\description{
Data for Exercise E, Chapter 03
}
\examples{
dse03e
}
\keyword{datasets}

