% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03ee}
\alias{dse03ee}
\title{Dataset for Exercise EE, Chapter 03}
\format{A \code{data.frame} with 23 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03ee
}
\description{
Dataset for Exercise EE, Chapter 03
}
\examples{
dse03ee
}
\keyword{datasets}

