% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dse03gg.R
\docType{data}
\name{dse03gg}
\alias{dse03gg}
\title{Dataset for Exercise GG, Chapter 03}
\format{A \code{data.frame} with 5 rows and 2 variables:
\describe{
\item{wage}{}
\item{turnover}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03gg
}
\description{
Dataset for Exercise GG, Chapter 03
}
\examples{
dse03gg
}
\keyword{datasets}

