% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03gg}
\alias{dse03gg}
\title{E03GG data}
\format{A data.frame with 5 rows and 2 variables:
\describe{
\item{wage}{}
\item{turnover}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03gg
}
\description{
Data for Exercise GG, Chapter 03
}
\examples{
dse03gg
}
\keyword{datasets}

