% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03h}
\alias{dse03h}
\title{Dataset for Exercise H, Chapter 03}
\format{A data.frame with 14 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03h
}
\description{
Dataset for Exercise H, Chapter 03
}
\examples{
dse03h
}
\keyword{datasets}

