% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03hh}
\alias{dse03hh}
\title{E03HH data}
\format{A data.frame with 29 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03hh
}
\description{
Dataset for Exercise HH, Chapter 03
}
\examples{
dse03hh
}
\keyword{datasets}

