% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03ii}
\alias{dse03ii}
\title{E03II data}
\format{A data.frame with 32 rows and 2 variables:
\describe{
\item{y}{}
\item{x}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03ii
}
\description{
Data for Exercise II, Chapter 03
}
\examples{
dse03ii
}
\keyword{datasets}

