% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse03ii.R
\docType{data}
\name{dse03ii}
\alias{dse03ii}
\title{Dataset for Exercise II, Chapter 03}
\format{A \code{data.frame} with 32 rows and 2 variables:
\describe{
\item{y}{}
\item{x}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03ii
}
\description{
Dataset for Exercise II, Chapter 03
}
\examples{
dse03ii
}
\keyword{datasets}

