% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03jj}
\alias{dse03jj}
\title{E03JJ data}
\format{A data.frame with 116 rows and 2 variables:
\describe{
\item{y}{}
\item{x}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03jj
}
\description{
Dataset for Exercise JJ, Chapter 03
}
\examples{
dse03jj
}
\keyword{datasets}

