% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03jj3}
\alias{dse03jj3}
\title{Dataset for Exercise JJ3, Chapter 03}
\format{A \code{data.frame} with 48 rows and 2 variables:
\describe{
\item{y}{}
\item{x}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03jj3
}
\description{
Dataset for Exercise JJ3, Chapter 03
}
\examples{
dse03jj3
}
\keyword{datasets}

