% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03k}
\alias{dse03k}
\title{E03K data}
\format{A data.frame with 34 rows and 2 variables:
\describe{
\item{y}{}
\item{x}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03k
}
\description{
Data for Exercise K, Chapter 03
}
\examples{
dse03k
}
\keyword{datasets}

