% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03ll}
\alias{dse03ll}
\title{Dataset for Exercise LL, Chapter 03}
\format{A data.frame with 13 rows and 2 variables:
\describe{
\item{y}{}
\item{x}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03ll
}
\description{
Dataset for Exercise LL, Chapter 03
}
\examples{
dse03ll
}
\keyword{datasets}

