% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03n}
\alias{dse03n}
\title{E03N data}
\format{A data.frame with 12 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03n
}
\description{
Dataset for Exercise N, Chapter 03
}
\examples{
dse03n
}
\keyword{datasets}

