% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03r}
\alias{dse03r}
\title{Dataset for Exercise R, Chapter 03}
\format{A \code{data.frame} with 43 rows and 2 variables:
\describe{
\item{t}{}
\item{a}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03r
}
\description{
Dataset for Exercise R, Chapter 03
}
\examples{
dse03r
}
\keyword{datasets}

