% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03v}
\alias{dse03v}
\title{E03V data}
\format{A data.frame with 50 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03v
}
\description{
Dataset for Exercise V, Chapter 03
}
\examples{
dse03v
}
\keyword{datasets}

