% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse03w.R
\docType{data}
\name{dse03w}
\alias{dse03w}
\title{Dataset for Exercise W, Chapter 03}
\format{
A \code{data.frame} with 14 rows and 2 variables:
\describe{
\item{year}{}
\item{price}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03w
}
\description{
Dataset for Exercise W, Chapter 03
}
\examples{
dse03w
}
\keyword{datasets}
