% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse03w}
\alias{dse03w}
\title{E03W data}
\format{A data.frame with 14 rows and 2 variables:
\describe{
\item{year}{}
\item{price}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03w
}
\description{
Dataset for Exercise W, Chapter 03
}
\examples{
dse03w
}
\keyword{datasets}

