% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse03z.R
\docType{data}
\name{dse03z}
\alias{dse03z}
\title{Dataset for Exercise Z, Chapter 03}
\format{A \code{data.frame} with 17 rows and 2 variables:
\describe{
\item{consumption}{}
\item{population}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse03z
}
\description{
Dataset for Exercise Z, Chapter 03
}
\examples{
dse03z
}
\keyword{datasets}

