% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse04d.R
\docType{data}
\name{dse04d}
\alias{dse04d}
\title{Dataset for Exercise D, Chapter 04}
\format{A \code{data.frame} with 14 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse04d
}
\description{
Dataset for Exercise D, Chapter 04
}
\examples{
dse04d
}
\keyword{datasets}

