% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse04e}
\alias{dse04e}
\title{Dataset for Exercise E, Chapter 04}
\format{A \code{data.frame} with 10 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse04e
}
\description{
Dataset for Exercise E, Chapter 04
}
\examples{
dse04e
}
\keyword{datasets}

