% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse04f}
\alias{dse04f}
\title{Dataset for Exercise F, Chapter 04}
\format{A \code{data.frame} with 10 rows and 3 variables:
\describe{
\item{x}{}
\item{y}{}
\item{xy}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse04f
}
\description{
Dataset for Exercise F, Chapter 04
}
\examples{
dse04f
}
\keyword{datasets}

