% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse05g}
\alias{dse05g}
\title{Dataset for Exercise G, Chapter 05}
\format{A data.frame with 9 rows and 6 variables:
\describe{
\item{row}{}
\item{y}{}
\item{dum1}{}
\item{dum2}{}
\item{speed}{}
\item{speed2}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse05g
}
\description{
Dataset for Exercise G, Chapter 05
}
\examples{
dse05g
}
\keyword{datasets}

