% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse05g}
\alias{dse05g}
\title{E05G data}
\format{A data.frame with 9 rows and 6 variables:
\describe{
\item{row}{}
\item{y}{}
\item{dum1}{}
\item{dum2}{}
\item{speed}{}
\item{speed2}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse05g
}
\description{
dse05g data
}
\examples{
dse05g
}
\keyword{datasets}

