% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dse06a.R
\docType{data}
\name{dse06a}
\alias{dse06a}
\title{Dataset for Exercise A, Chapter 06}
\format{A \code{data.frame} with 11 rows and 4 variables:
\describe{
\item{x0}{}
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06a
}
\description{
Dataset for Exercise A, Chapter 06
}
\examples{
dse06a
}
\keyword{datasets}

