% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse06a}
\alias{dse06a}
\title{E06A data}
\format{A data.frame with 11 rows and 4 variables:
\describe{
\item{x0}{}
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06a
}
\description{
Dataset for Exercise A, Chapter 06
}
\examples{
dse06a
}
\keyword{datasets}

