% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse06b}
\alias{dse06b}
\title{E06B data}
\format{A data.frame with 12 rows and 3 variables:
\describe{
\item{x}{}
\item{y}{}
\item{z}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06b
}
\description{
Dataset for Exercise B, Chapter 06
}
\examples{
dse06b
}
\keyword{datasets}

