% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse06b.R
\docType{data}
\name{dse06b}
\alias{dse06b}
\title{Dataset for Exercise B, Chapter 06}
\format{A \code{data.frame} with 12 rows and 3 variables:
\describe{
\item{x}{}
\item{y}{}
\item{z}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06b
}
\description{
Dataset for Exercise B, Chapter 06
}
\examples{
dse06b
}
\keyword{datasets}

