% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse06c}
\alias{dse06c}
\title{Dataset for Exercise C, Chapter 06}
\format{A \code{data.frame} with 15 rows and 3 variables:
\describe{
\item{y}{}
\item{x1}{}
\item{x2}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06c
}
\description{
Dataset for Exercise C, Chapter 06
}
\examples{
dse06c
}
\keyword{datasets}

