% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse06d}
\alias{dse06d}
\title{E06D data}
\format{A data.frame with 8 rows and 3 variables:
\describe{
\item{y}{}
\item{x1}{}
\item{x2}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06d
}
\description{
Dataset for Exercise D, Chapter 06
}
\examples{
dse06d
}
\keyword{datasets}

