% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dse06d.R
\docType{data}
\name{dse06d}
\alias{dse06d}
\title{Dataset for Exercise D, Chapter 06}
\format{A \code{data.frame} with 8 rows and 3 variables:
\describe{
\item{y}{}
\item{x1}{}
\item{x2}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06d
}
\description{
Dataset for Exercise D, Chapter 06
}
\examples{
dse06d
}
\keyword{datasets}

