% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse06g}
\alias{dse06g}
\title{E06G data}
\format{A data.frame with 17 rows and 3 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06g
}
\description{
Dataset for Exercise G, Chapter 06
}
\examples{
dse06g
}
\keyword{datasets}

