% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse06h}
\alias{dse06h}
\title{E06H data}
\format{A data.frame with 13 rows and 3 variables:
\describe{
\item{y}{}
\item{x}{}
\item{z}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06h
}
\description{
Dataset for Exercise H, Chapter 06
}
\examples{
dse06h
}
\keyword{datasets}

