% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse06j.R
\docType{data}
\name{dse06j}
\alias{dse06j}
\title{Dataset for Exercise J, Chapter 06}
\format{A \code{data.frame} with 10 rows and 3 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06j
}
\description{
Dataset for Exercise J, Chapter 06
}
\examples{
dse06j
}
\keyword{datasets}

