% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse06k}
\alias{dse06k}
\title{E06K data}
\format{A data.frame with 5 rows and 3 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06k
}
\description{
Data for Exercise K, Chapter 06
}
\examples{
dse06k
}
\keyword{datasets}

