% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dse06l.R
\docType{data}
\name{dse06l}
\alias{dse06l}
\title{Dataset for Exercise L, Chapter 06}
\format{A \code{data.frame} with 9 rows and 4 variables:
\describe{
\item{batchn}{}
\item{t}{}
\item{c}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06l
}
\description{
Dataset for Exercise L, Chapter 06
}
\examples{
dse06l
}
\keyword{datasets}

