% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse06z.R
\docType{data}
\name{dse06z}
\alias{dse06z}
\title{Dataset for Exercise Z, Chapter 06}
\format{A \code{data.frame} with 19 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse06z
}
\description{
Dataset for Exercise Z, Chapter 06
}
\examples{
dse06z
}
\keyword{datasets}

