% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse07b.R
\docType{data}
\name{dse07b}
\alias{dse07b}
\title{Dataset for Exercise B, Chapter 07}
\format{A \code{data.frame} with 48 rows and 4 variables:
\describe{
\item{time}{}
\item{set1}{}
\item{set2}{}
\item{set3}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse07b
}
\description{
Dataset for Exercise B, Chapter 07
}
\examples{
dse07b
}
\keyword{datasets}

