% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse07b}
\alias{dse07b}
\title{Dataset for Exercise B, Chapter 07}
\format{A data.frame with 48 rows and 3 variables:
\describe{
\item{set1}{}
\item{set2}{}
\item{set3}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse07b
}
\description{
Dataset for Exercise B, Chapter 07
}
\examples{
dse07b
}
\keyword{datasets}

