% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse07c}
\alias{dse07c}
\title{E07C data}
\format{A data.frame with 50 rows and 7 variables:
\describe{
\item{y}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{x6}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse07c
}
\description{
Data for Exercise C, Chapter 07
}
\examples{
dse07c
}
\keyword{datasets}

