% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse07c.R
\docType{data}
\name{dse07c}
\alias{dse07c}
\title{Dataset for Exercise C, Chapter 07}
\format{A \code{data.frame} with 50 rows and 7 variables:
\describe{
\item{y}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{x6}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse07c
}
\description{
Dataset for Exercise C, Chapter 07
}
\examples{
dse07c
}
\keyword{datasets}

