% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse08b}
\alias{dse08b}
\title{Dataset for Exercise B, Chapter 08}
\format{A \code{data.frame} with 5 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse08b
}
\description{
Dataset for Exercise B, Chapter 08
}
\examples{
dse08b
}
\keyword{datasets}

