% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse09b}
\alias{dse09b}
\title{E09B data}
\format{A data.frame with 5 rows and 3 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse09b
}
\description{
Data for Exercise B, Chapter 09
}
\examples{
dse09b
}
\keyword{datasets}

