% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse12a}
\alias{dse12a}
\title{Dataset for Exercise A, Chapter 12}
\format{A \code{data.frame} with 18 rows and 5 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse12a
}
\description{
Dataset for Exercise A, Chapter 12
}
\examples{
dse12a
}
\keyword{datasets}

