% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dse12b.R
\docType{data}
\name{dse12b}
\alias{dse12b}
\title{Dataset for Exercise B, Chapter 12}
\format{A \code{data.frame} with 20 rows and 4 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse12b
}
\description{
Dataset for Exercise B, Chapter 12
}
\examples{
dse12b
}
\keyword{datasets}

