% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse12c}
\alias{dse12c}
\title{Dataset for Exercise C, Chapter 12}
\format{A data.frame with 24 rows and 2 variables:
\describe{
\item{t}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse12c
}
\description{
Dataset for Exercise C, Chapter 12
}
\examples{
dse12c
}
\keyword{datasets}

