% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse12e}
\alias{dse12e}
\title{E12E data}
\format{A data.frame with 28 rows and 5 variables:
\describe{
\item{fat}{}
\item{flour}{}
\item{water}{}
\item{rpm}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse12e
}
\description{
Data for Exercise E, Chapter 12
}
\examples{
dse12e
}
\keyword{datasets}

