% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse12h}
\alias{dse12h}
\title{E12H data}
\format{A data.frame with 27 rows and 4 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse12h
}
\description{
Data for Exercise H, Chapter 12
}
\examples{
dse12h
}
\keyword{datasets}

