% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse13a}
\alias{dse13a}
\title{E13A data}
\format{A data.frame with 6 rows and 3 variables:
\describe{
\item{id}{}
\item{x}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse13a
}
\description{
dse13a data
}
\examples{
dse13a
}
\keyword{datasets}

