% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse13c.R
\docType{data}
\name{dse13c}
\alias{dse13c}
\title{Dataset for Exercise C, Chapter 13}
\format{A \code{data.frame} with 11 rows and 6 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{y1}{}
\item{y2}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse13c
}
\description{
Dataset for Exercise C, Chapter 13
}
\examples{
dse13c
}
\keyword{datasets}

