% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse13d}
\alias{dse13d}
\title{E13D data}
\format{A data.frame with 24 rows and 4 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse13d
}
\description{
Dataset for Exercise D, Chapter 13
}
\examples{
dse13d
}
\keyword{datasets}

