% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse13d.R
\docType{data}
\name{dse13d}
\alias{dse13d}
\title{Dataset for Exercise D, Chapter 13}
\format{
A \code{data.frame} with 24 rows and 4 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse13d
}
\description{
Dataset for Exercise D, Chapter 13
}
\examples{
dse13d
}
\keyword{datasets}
