% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse13e}
\alias{dse13e}
\title{E13E data}
\format{A data.frame with 43 rows and 2 variables:
\describe{
\item{t}{}
\item{m}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse13e
}
\description{
Data for Exercise E, Chapter 13
}
\examples{
dse13e
}
\keyword{datasets}

