% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dse13f.R
\docType{data}
\name{dse13f}
\alias{dse13f}
\title{Dataset for Exercise F, Chapter 13}
\format{A \code{data.frame} with 16 rows and 3 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse13f
}
\description{
Dataset for Exercise F, Chapter 13
}
\examples{
dse13f
}
\keyword{datasets}

