% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse13f}
\alias{dse13f}
\title{E13F data}
\format{A data.frame with 16 rows and 4 variables:
\describe{
\item{id}{}
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse13f
}
\description{
dse13f data
}
\examples{
dse13f
}
\keyword{datasets}

