% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse13g}
\alias{dse13g}
\title{Dataset for Exercise G, Chapter 13}
\format{A data.frame with 35 rows and 4 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse13g
}
\description{
Dataset for Exercise G, Chapter 13
}
\examples{
dse13g
}
\keyword{datasets}

