% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse13h.R
\docType{data}
\name{dse13h}
\alias{dse13h}
\title{Dataset for Exercise H, Chapter 13}
\format{A \code{data.frame} with 8 rows and 5 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse13h
}
\description{
Dataset for Exercise H, Chapter 13
}
\examples{
dse13h
}
\keyword{datasets}

