% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse14l}
\alias{dse14l}
\title{E14L data}
\format{A data.frame with 8 rows and 4 variables:
\describe{
\item{id}{}
\item{x}{}
\item{dummy}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse14l
}
\description{
dse14l data
}
\examples{
dse14l
}
\keyword{datasets}

