% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse14l}
\alias{dse14l}
\title{Dataset for Exercise L, Chapter 14}
\format{A \code{data.frame} with 8 rows and 3 variables:
\describe{
\item{x}{}
\item{dummy}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse14l
}
\description{
Dataset for Exercise L, Chapter 14
}
\examples{
dse14l
}
\keyword{datasets}

