% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse14r}
\alias{dse14r}
\title{E14R data}
\format{A data.frame with 48 rows and 5 variables:
\describe{
\item{id}{}
\item{boot}{}
\item{subject}{}
\item{temperature}{}
\item{drop}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse14r
}
\description{
dse14r data
}
\examples{
dse14r
}
\keyword{datasets}

