% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse14s}
\alias{dse14s}
\title{E14S data}
\format{A data.frame with 8 rows and 4 variables:
\describe{
\item{x}{}
\item{z}{}
\item{zx}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse14s
}
\description{
Dataset for Exercise S, Chapter 14
}
\examples{
dse14s
}
\keyword{datasets}

