% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse14t}
\alias{dse14t}
\title{Dataset for Exercise T, Chapter 14}
\format{A \code{data.frame} with 15 rows and 4 variables:
\describe{
\item{x}{}
\item{z}{}
\item{xz}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse14t
}
\description{
Dataset for Exercise T, Chapter 14
}
\examples{
dse14t
}
\keyword{datasets}

