% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse14u}
\alias{dse14u}
\title{E14U data}
\format{A data.frame with 15 rows and 4 variables:
\describe{
\item{z1}{}
\item{z2}{}
\item{z3}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse14u
}
\description{
Dataset for Exercise U, Chapter 14
}
\examples{
dse14u
}
\keyword{datasets}

