% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse15a.R
\docType{data}
\name{dse15a}
\alias{dse15a}
\title{Dataset for Exercise A, Chapter 15}
\format{
A \code{data.frame} with 17 rows and 6 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{y}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse15a
}
\description{
Dataset for Exercise A, Chapter 15
}
\examples{
dse15a
}
\keyword{datasets}
