% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dse15b.R
\docType{data}
\name{dse15b}
\alias{dse15b}
\title{Dataset for Exercise B, Chapter 15}
\format{A \code{data.frame} with 9 rows and 4 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse15b
}
\description{
Dataset for Exercise B, Chapter 15
}
\examples{
dse15b
}
\keyword{datasets}

