% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse15c}
\alias{dse15c}
\title{Dataset for Exercise C, Chapter 15}
\format{A data.frame with 20 rows and 6 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse15c
}
\description{
Dataset for Exercise C, Chapter 15
}
\examples{
dse15c
}
\keyword{datasets}

