% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse15j.R
\docType{data}
\name{dse15j}
\alias{dse15j}
\title{Dataset for Exercise J, Chapter 15}
\format{
A \code{data.frame} with 16 rows and 7 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{x6}{}
\item{y}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse15j
}
\description{
Dataset for Exercise J, Chapter 15
}
\examples{
dse15j
}
\keyword{datasets}
