% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse15l}
\alias{dse15l}
\title{E15L data}
\format{A data.frame with 15 rows and 9 variables:
\describe{
\item{id}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{x6}{}
\item{x7}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse15l
}
\description{
dse15l data
}
\examples{
dse15l
}
\keyword{datasets}

