% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse15n}
\alias{dse15n}
\title{Dataset for Exercise N, Chapter 15}
\format{A \code{data.frame} with 21 rows and 4 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse15n
}
\description{
Dataset for Exercise N, Chapter 15
}
\examples{
dse15n
}
\keyword{datasets}

