% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse16a}
\alias{dse16a}
\title{Dataset for Exercise A, Chapter 16}
\format{A data.frame with 5 rows and 4 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse16a
}
\description{
Dataset for Exercise A, Chapter 16
}
\examples{
dse16a
}
\keyword{datasets}

