% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse16c.R
\docType{data}
\name{dse16c}
\alias{dse16c}
\title{Dataset for Exercise C, Chapter 16}
\format{
A \code{data.frame} with 7 rows and 3 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{y}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse16c
}
\description{
Dataset for Exercise C, Chapter 16
}
\examples{
dse16c
}
\keyword{datasets}
