% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse16d}
\alias{dse16d}
\title{Dataset for Exercise D, Chapter 16}
\format{A \code{data.frame} with 5 rows and 3 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse16d
}
\description{
Dataset for Exercise D, Chapter 16
}
\examples{
dse16d
}
\keyword{datasets}

