% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse17b.R
\docType{data}
\name{dse17b}
\alias{dse17b}
\title{Dataset for Exercise B, Chapter 17}
\format{
A \code{data.frame} with 8 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse17b
}
\description{
Dataset for Exercise B, Chapter 17
}
\examples{
dse17b
}
\keyword{datasets}
