% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse19d}
\alias{dse19d}
\title{Dataset for Exercise D, Chapter 19}
\format{A data.frame with 8 rows and 4 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse19d
}
\description{
Dataset for Exercise D, Chapter 19
}
\examples{
dse19d
}
\keyword{datasets}

