% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse22e.R
\docType{data}
\name{dse22e}
\alias{dse22e}
\title{Dataset for Exercise E, Chapter 22}
\format{
A \code{data.frame} with 9 rows and 7 variables:
\describe{
\item{y}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{x6}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse22e
}
\description{
Dataset for Exercise E, Chapter 22
}
\examples{
dse22e
}
\keyword{datasets}
