% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse22f}
\alias{dse22f}
\title{E22F data}
\format{A data.frame with 12 rows and 7 variables:
\describe{
\item{y}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
\item{x4}{}
\item{x5}{}
\item{x6}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse22f
}
\description{
Data for Exercise F, Chapter 22
}
\examples{
dse22f
}
\keyword{datasets}

