% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse22g}
\alias{dse22g}
\title{E22G data}
\format{A data.frame with 24 rows and 6 variables:
\describe{
\item{age}{}
\item{n}{}
\item{weight}{}
\item{sd}{}
\item{z1}{}
\item{z2}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse22g
}
\description{
Data for Exercise G, Chapter 22
}
\examples{
dse22g
}
\keyword{datasets}

