% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dse22g.R
\docType{data}
\name{dse22g}
\alias{dse22g}
\title{Dataset for Exercise G, Chapter 22}
\format{A \code{data.frame} with 24 rows and 5 variables:
\describe{
\item{age}{}
\item{n}{}
\item{weight}{}
\item{sd}{}
\item{category}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse22g
}
\description{
Dataset for Exercise G, Chapter 22
}
\examples{
dse22g
}
\keyword{datasets}

