% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse24a.R
\docType{data}
\name{dse24a}
\alias{dse24a}
\title{Dataset for Exercise A, Chapter 24}
\format{
A \code{data.frame} with 3 rows and 2 variables:
\describe{
\item{t}{}
\item{y}{}
}
}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24a
}
\description{
Dataset for Exercise A, Chapter 24
}
\examples{
dse24a
}
\keyword{datasets}
