% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dse24b.R
\docType{data}
\name{dse24b}
\alias{dse24b}
\title{Dataset for Exercise B, Chapter 24}
\format{A \code{data.frame} with 12 rows and 2 variables:
\describe{
\item{t}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24b
}
\description{
Dataset for Exercise B, Chapter 24
}
\examples{
dse24b
}
\keyword{datasets}

