% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24c}
\alias{dse24c}
\title{Dataset for Exercise C, Chapter 24}
\format{A data.frame with 4 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24c
}
\description{
Dataset for Exercise C, Chapter 24
}
\examples{
dse24c
}
\keyword{datasets}

