% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24d}
\alias{dse24d}
\title{E24D data}
\format{A data.frame with 5 rows and 2 variables:
\describe{
\item{x}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24d
}
\description{
Data for Exercise D, Chapter 24
}
\examples{
dse24d
}
\keyword{datasets}

