% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24e}
\alias{dse24e}
\title{E24E data}
\format{A data.frame with 14 rows and 3 variables:
\describe{
\item{id}{}
\item{t}{}
\item{y}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24e
}
\description{
dse24e data
}
\examples{
dse24e
}
\keyword{datasets}

