% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24h}
\alias{dse24h}
\title{E24H data}
\format{A data.frame with 38 rows and 3 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24h
}
\description{
Dataset for Exercise H, Chapter 24
}
\examples{
dse24h
}
\keyword{datasets}

