% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dse24i.R
\docType{data}
\name{dse24i}
\alias{dse24i}
\title{Dataset for Exercise I, Chapter 24}
\format{A \code{data.frame} with 8 rows and 3 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24i
}
\description{
Dataset for Exercise I, Chapter 24
}
\examples{
dse24i
}
\keyword{datasets}

