% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24i}
\alias{dse24i}
\title{E24I data}
\format{A data.frame with 8 rows and 3 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24i
}
\description{
Data for Exercise I, Chapter 24
}
\examples{
dse24i
}
\keyword{datasets}

