% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24j}
\alias{dse24j}
\title{Dataset for Exercise J, Chapter 24}
\format{A \code{data.frame} with 13 rows and 3 variables:
\describe{
\item{x1}{}
\item{x2}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24j
}
\description{
Dataset for Exercise J, Chapter 24
}
\examples{
dse24j
}
\keyword{datasets}

