% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24k}
\alias{dse24k}
\title{E24K data}
\format{A data.frame with 5 rows and 2 variables:
\describe{
\item{t}{}
\item{y}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24k
}
\description{
Data for Exercise K, Chapter 24
}
\examples{
dse24k
}
\keyword{datasets}

