% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dse24m.R
\docType{data}
\name{dse24m}
\alias{dse24m}
\title{Dataset for Exercise M, Chapter 24}
\format{A \code{data.frame} with 43 rows and 2 variables:
\describe{
\item{t}{}
\item{m}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24m
}
\description{
Dataset for Exercise M, Chapter 24
}
\examples{
dse24m
}
\keyword{datasets}

