% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24n}
\alias{dse24n}
\title{E24N data}
\format{A data.frame with 7 rows and 6 variables:
\describe{
\item{t}{}
\item{w1}{}
\item{w2}{}
\item{w3}{}
\item{w4}{}
\item{w5}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24n
}
\description{
Dataset for Exercise N, Chapter 24
}
\examples{
dse24n
}
\keyword{datasets}

