% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24p}
\alias{dse24p}
\title{E24P data}
\format{A data.frame with 5 rows and 3 variables:
\describe{
\item{x}{}
\item{y}{}
\item{z}{}
}}
\source{
Draper, N.R., Smith, H., (1998) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24p
}
\description{
Dataset for Exercise P, Chapter 24
}
\examples{
dse24p
}
\keyword{datasets}

