% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dse24p}
\alias{dse24p}
\title{E24P data}
\format{A data.frame with 5 rows and 4 variables:
\describe{
\item{id}{}
\item{x}{}
\item{y}{}
\item{z}{}
}}
\source{
Draper, X.Y., Smith, X.Y., (1999) Applied Regression Analyis, 3rd ed., New York: Wiley
}
\usage{
dse24p
}
\description{
dse24p data
}
\examples{
dse24p
}
\keyword{datasets}

